+++ lib/feature/storage_operation/placement_writing_off/placement_writeing_off_repository/model/cell_model.dart
@@ -31,20 +31,19 @@
 // }
 import 'package:equatable/equatable.dart';
 
-class Cell extends Equatable{
+class Cell extends Equatable {
   final List<CellData> cell;
   final int zone;
 
   const Cell({required this.cell, required this.zone});
-  
+
   @override
   List<Object?> get props => [cell, zone];
 
   static const empty = Cell(cell: [CellData.empty], zone: 1);
-
 }
 
-class CellData extends Equatable{
+class CellData extends Equatable {
   final String name;
   final dynamic quantity;
   final String nameCell;
@@ -52,7 +51,7 @@ class CellData extends Equatable{
   final List<String> barcodes;
   final String article;
 
-  final int status;
+  final double status;
 
   const CellData(
       {required this.name,
@@ -62,9 +61,16 @@ class CellData extends Equatable{
       required this.article,
       required this.barcodes,
       required this.status});
-        @override
-  List<Object?> get props => [name, quantity, barcodes,  status, nameCell, codeCell,article ];
+  @override
+  List<Object?> get props =>
+      [name, quantity, barcodes, status, nameCell, codeCell, article];
 
-    static const empty = CellData(name: '', quantity: 0, nameCell: '', codeCell: '', article: '', barcodes: [], status: 1);
-
+  static const empty = CellData(
+      name: '',
+      quantity: 0,
+      nameCell: '',
+      codeCell: '',
+      article: '',
+      barcodes: [],
+      status: 1);
 }
